// core/types/enums/RecipeID.ts
/**
 * Если в конце число, то добавляется знак _ или для каждого рецепта можно делать читаемое осмысленное название.
 */
export enum RecipeID {
    // Wood = 'Wood',
    WoodBoards = 'WoodBoards',

    // Coal = 'Coal',

    // IronOre = 'IronOre',
    IronIngot = 'IronIngot',
    // CopperOre = 'CopperOre',
    CopperIngot = 'CopperIngot',
    // TinOre = 'TinOre',
    TinIngot = 'TinIngot',
    BronzeIngot = 'BronzeIngot',
    // GoldOre = 'GoldOre',
    GoldIngot = 'GoldIngot',

    // Herb01 = 'Herb01',
    // Herb02 = 'Herb02',
    // Herb03 = 'Herb03',

    // Cotton = 'Cotton',
    CottonThread = 'CottonThread',
    CottonFabric = 'CottonFabric',

    Leather01 = 'Leather01',
    // Leather01FromRabbitSkin = 'Leather01FromRabbitSkin',
    // Leather01FromBoarSkin = 'Leather01FromBoarSkin',
    // Leather01FromFoxSkin = 'Leather01FromFoxSkin',
    // Leather01FromDearSkin = 'Leather01FromDearSkin',
    // Leather01FromWolfSkin = 'Leather01FromWolfSkin',
    // Leather01FromBearSkin = 'Leather01FromBearSkin',

    // BoarMeat = 'BoarMeat',
    // BearMeat = 'BearMeat',

    // MagicResources01 = 'MagicResources01',
    MagicMaterial01 = 'MagicMaterial01',

    HealthPotion01 = 'HealthPotion01',
    MagicPotion02 = 'MagicPotion02',

    PlateHelmet01 = 'PlateHelmet01',
    PlateHelmet02 = 'PlateHelmet02',
    PlateShoulders01 = 'PlateShoulders01',
    PlateBreastplate01 = 'PlateBreastplate01',
    PlateBracer01 = 'PlateBracer01',
    PlateGloves01 = 'PlateGloves01',
    PlateBelt01 = 'PlateBelt01',
    PlatePants01 = 'PlatePants01',
    PlateBoots01 = 'PlateBoots01',

    LeatherHelmet01 = 'LeatherHelmet01',
    LeatherShoulders01 = 'LeatherShoulders01',
    LeatherBreastplate01 = 'LeatherBreastplate01',
    LeatherBracer01 = 'LeatherBracer01',
    LeatherGloves01 = 'LeatherGloves01',
    LeatherBelt01 = 'LeatherBelt01',
    LeatherPants01 = 'LeatherPants01',
    LeatherBoots01 = 'LeatherBoots01',

    ClothHelmet01 = 'ClothHelmet01',
    ClothShoulders_01 = 'ClothShoulders_01',
    ClothBreastplate01 = 'ClothBreastplate01',
    ClothBracer01 = 'ClothBracer01',
    ClothGloves01 = 'ClothGloves01',
    ClothBelt01 = 'ClothBelt01',
    ClothPants01 = 'ClothPants01',
    ClothBoots01 = 'ClothBoots01',

    Shield01 = 'Shield01',
    Shield02 = 'Shield02',

    Amulet01 = 'Amulet01',
    Amulet02 = 'Amulet02',
    Amulet03 = 'Amulet03',
    Amulet04 = 'Amulet04',
    Amulet05 = 'Amulet05',
    Ring01 = 'Ring01',
    Ring02 = 'Ring02',
    Ring03 = 'Ring03',
    Ring04 = 'Ring04',
    Ring05 = 'Ring05',
    Trinket01 = 'Trinket01',
    Trinket02 = 'Trinket02',
    Trinket03 = 'Trinket03',
    Trinket04 = 'Trinket04',
    Trinket05 = 'Trinket05',

    OneHandedSword01 = 'OneHandedSword01',
    TwoHandedSword01 = 'TwoHandedSword01',
    Dagger01 = 'Dagger01',
    OneHandedAxe01 = 'OneHandedAxe01',
    TwoHandedAxe01 = 'TwoHandedAxe01',
    Staff01 = 'Staff01',
    Wand01 = 'Wand01',
    Revolver01 = 'Revolver01',
}